ARREGLAR SALE Y RENT

+++ TrabajoPW/src/main/java/pe/edu/upc/repository/IRentRepository.java
@@ -1,10 +1,16 @@
 package pe.edu.upc.repository;
 
+
+
 import org.springframework.data.jpa.repository.JpaRepository;
+
 import org.springframework.stereotype.Repository;
+
 
 import pe.edu.upc.entity.Rent;
 @Repository
-public interface IRentRepository extends JpaRepository<Rent, Integer>{
+public interface IRentRepository extends JpaRepository<Rent, Long>{
+
+	
 
 }
